Extract local notification creation from UploadManager to NotificationManager.
