+++ Core/Core/SubmitAssignmentExtension/Model/AttachmentSubmissionService.swift
@@ -38,7 +38,7 @@ public class AttachmentSubmissionService {
         ProcessInfo.processInfo.performExpiringActivity(withReason: "get upload targets") { expired in
             if expired {
                 Analytics.shared.logError("error_performing_background_activity")
-                self.uploadManager.sendFailedNotification()
+                self.uploadManager.notificationManager.sendFailedNotification()
                 return
             }
             self.uploadManager.viewContext.perform {
@@ -58,7 +58,7 @@ public class AttachmentSubmissionService {
                 }
             }
             if error != nil {
-                self.uploadManager.sendFailedNotification()
+                self.uploadManager.notificationManager.sendFailedNotification()
             }
             semaphore.wait()
             callback()
